Clean-up - File renaming and README updates:
The repository was a mess, and it needed cleaned up.

I removed some playbooks that will be turned into roles and hosted in
their own repositories, so now this repository is only for the initial
installation of ansible-controller (bootstrapping), and initial
configurations of hosts.

I tried to clean up the README, but it is still way too long. Will
probably move the security context and other things to a separate file
down the road.